working implementation for macos

+++ ping_pong_camera.rb
@@ -4,25 +4,27 @@ require_relative 'video_splicer.rb'
 class Supervisor
   def initialize
     @recorder = Recorder.new
+    @is_joining = false
+    File.delete('/tmp/event.log') unless !File.exists?('/tmp/event.log')
   end
 
   def record_video
     @is_recording = @recorder.record
-    sleep 15
-    @is_recording = false
-    record_video
   end
 
   def listen_for_messages
     loop do
       begin
         contents = File.read('/tmp/event.log')
-	VideoSplicer.join
-	File.delete('/tmp/event.log')
+        if !@is_joining
+          @is_joining = true
+          VideoSplicer.join
+	  @is_joining = false
+          File.delete('/tmp/event.log') unless !File.exists?('/tmp/event.log')
+	end
       rescue
       ensure
-        sleep 1
-        listen_for_messages
+        sleep 5
       end
     end
   end
@@ -34,4 +36,8 @@ class Supervisor
 end
 
 supervisor = Supervisor.new
-supervisor.start
+
+begin
+  supervisor.start
+rescue Interrupt => e
+end
